Fix reinstallations

+++ setup.sh
@@ -39,15 +39,16 @@ owodir="$HOME/.config/owo"
 
 if [ ! -d $owodir ]; then
 	mkdir $owodir
-	cp -r $scriptdir/* $owodir
 fi
+cp -r $scriptdir/* $owodir
 
 # Give directory ownership to the actual user
 chown -R $(whoami | awk '{print $1}') $owodir
 
 # Create a symbolic link to /usr/local/bin
+if [ ! -f /usr/local/bin/owo ]; then
 sudo ln -s $owodir/script.sh /usr/local/bin/owo
-
+fi
 function is_mac() {
 	uname | grep -q "Darwin"
 }
